use View::call to render using presenter object, instead of instead of introducing a new Presenter.
