+++ src/Helper/ResponderBuilder.php
@@ -15,24 +15,25 @@ class ResponderBuilder
 {
     /**
      * build responder object based on $view (ViewerInterface) and
-     * $error( ErrorViewInterface) services. 
-     * 
+     * $error( ErrorViewInterface) services.
+     *
      * @param ViewerInterface    $view
      * @param ErrorViewInterface $error
      * @param null|string        $content_view
+     * @param null|callable      $resolver
      * @return Responder
      */
     public static function withServices(
         ViewerInterface $view,
         ErrorViewInterface $error,
-        $content_view = null
+        $content_view = null,
+        $resolver = null
     ) {
         $self = new Responder(
-            new View($view, $content_view),
+            new View($view, $content_view, $resolver),
             new Redirect(),
             new Error($error),
-            new ViewData(),
-            new Presenter()
+            new ViewData()
         );
 
         return $self;
@@ -40,17 +41,18 @@ class ResponderBuilder
 
     /**
      * build responder object based on $view (ViewerInterface) object
-     * and options for errors. 
-     * 
+     * and options for errors.
+     *
      * @param ViewerInterface $view
      * @param array           $errorOption
      * @param string|null     $content_view
+     * @param null|callable   $resolver
      * @return Responder
      */
-    public static function withView(ViewerInterface $view, $errorOption = [], $content_view = null)
+    public static function withView(ViewerInterface $view, $errorOption = [], $content_view = null, $resolver = null)
     {
         $errors = ErrorView::forge($view, $errorOption);
 
-        return self::withServices($view, $errors, $content_view);
+        return self::withServices($view, $errors, $content_view, $resolver);
     }
 }
