+++ tests/Responder/ViewTest.php
@@ -8,6 +8,7 @@ use Tuum\Respond\Responder;
 use Tuum\Respond\Responder\View;
 use Tuum\Respond\Service\SessionStorage;
 use Tuum\Respond\Service\SessionStorageInterface;
+use Tuum\Respond\Service\ViewerInterface;
 use Zend\Diactoros\Response;
 
 class ViewTest extends \PHPUnit_Framework_TestCase
@@ -95,5 +96,60 @@ class ViewTest extends \PHPUnit_Framework_TestCase
         $response = $this->view->asFileContents($fp, 'mime/test');
         $this->assertSame('test/resource', $response->getBody()->__toString());
         $this->assertEquals('mime/test', $response->getHeader('Content-Type')[0]);
+    }
+
+    /**
+     * @test
+     */
+    function executes_callable()
+    {
+        $string = $this->view->call(function() { return 'tested: closure';});
+        $this->assertEquals('tested: closure', $string);
+    }
+
+    /**
+     * @test
+     */
+    function executes_ViewerInterface_object()
+    {
+        $stub = $this->getMockBuilder(ViewerInterface::class)->getMock();
+        $stub->method('withView')->willReturn('tested: mock');
+        $string = $this->view->call($stub);
+        $this->assertEquals('tested: mock', $string);
+    }
+
+    /**
+     * @test
+     * @expectedException \BadMethodCallException
+     */
+    function execute_string_without_resolver_throws_exception()
+    {
+        $this->view->call('bad-input');
+    }
+
+    /**
+     * @test
+     * @expectedException \InvalidArgumentException
+     */
+    function returning_non_callable_from_resolver_throws_exception()
+    {
+        $view = new View(new LocalView(), null, function() {return 'not-a-callable';});
+        $view->call('bad-input');
+    }
+
+    /**
+     * @test
+     */
+    function execute_using_resolver()
+    {
+        $view   = new View(
+            new LocalView(), null,
+            function() {
+                return function() {
+                    return 'tested: resolver';
+                };
+            });
+        $string = $view->call('any-string');
+        $this->assertEquals('tested: resolver', $string);
     }
 }
